FSD functionality removed
FSD functionality required messy links to an on-machine install of
sklearn. Really needs to be fully reimplemented in a separate library,
so has been removed from this commit onwards. The last commit prior to
this one had the FSD functionality (using constructor argument
mt_type='local') as used in the thesis.